+++ consensus/dummy/tests/backend.rs
@@ -4,7 +4,7 @@ extern crate ekiden_consensus_dummy;
 
 use std::sync::Arc;
 
-use ekiden_common::futures::{future, BoxFuture, Future, Stream};
+use ekiden_common::futures::{cpupool, future, Future, Stream};
 use ekiden_consensus_base::ConsensusBackend;
 use ekiden_consensus_base::test::create_computation_group;
 use ekiden_consensus_dummy::DummyConsensusBackend;
@@ -18,11 +18,17 @@ fn test_dummy_backend_two_rounds() {
         computation_group.iter().map(|n| n.get_public()).collect(),
     ));
 
+    let mut pool = cpupool::CpuPool::new(4);
+
+    // Start backend.
+    backend.start(&mut pool);
+
     // Start all nodes.
-    let mut node_tasks: Vec<BoxFuture<()>> = computation_group
+    let mut tasks = vec![];
+    tasks.append(&mut computation_group
         .iter()
         .map(|n| n.start(backend.clone()))
-        .collect();
+        .collect());
 
     // Send compute requests to all nodes.
     for ref node in computation_group.iter() {
@@ -46,6 +52,9 @@ fn test_dummy_backend_two_rounds() {
                 for ref node in computation_group.iter() {
                     node.shutdown();
                 }
+
+                let backend = backend.clone();
+                backend.shutdown();
             }
             round => panic!("incorrect round number: {}", round),
         }
@@ -53,8 +62,8 @@ fn test_dummy_backend_two_rounds() {
         Ok(())
     });
 
-    node_tasks.push(Box::new(wait_rounds));
+    tasks.push(Box::new(wait_rounds));
 
     // Wait for all tasks to finish.
-    future::join_all(node_tasks).wait().unwrap();
+    pool.spawn(future::join_all(tasks)).wait().unwrap();
 }